Show correct actual type in error message for unpacked lists.
Summary:
Problem: When passing `*list_of_string` to `*args: *Tuple[int, ...]`, we were showing "actual" type as `List[str]` instead of `str`.

Pass the item type for `*list_of_string` and use that in the error message.

That was the default behavior for `*args: int`. We were doing something different for `*args: *Tuple[<...>]`.

Reviewed By: stroxler

Differential Revision: D33370047

fbshipit-source-id: 22283d12113dde5a8f11ae9e10a67eff178d27c7