Check multiple unpacked arguments against `*args: *Tuple[<...>]`.
Summary:
Problem: We want to check multiple unpacked arguments: `*xs, 42, *ys` against an expected type `*args: *Tuple[int, *Tuple[int, ...], str]`.

Right now, this is not supported at all.

If there are multiple variadic tuples, we give up. But if there are multiple unbounded tuples, we basically preserve the concrete prefix and suffix and coalesce the middle elements into a single unbounded tuple type. By taking the union, we ensure soundness even if one of the tuples is empty at runtime.

Otherwise, because we don't know the length of an unbounded tuple, we might wrongly accept `*list_int, *list_string` for `*args: *Tuple[int, int, *Tuple[str, ...]]` and the `list_int` could be be empty at runtime, leading to crashes.

Reviewed By: stroxler

Differential Revision: D33370046

fbshipit-source-id: dea629e79e82f11b82a418d051f916a00c57938f