Temporarily remove test failing on Mac
Summary:
It works on Linux but not on Mac it looks like. Not sure why, but disabling temporarily while I investigate.

Interestingly, the test in `test_handle_query_with_build_system` for `ModulesOfPath` does pass on Mac; which is pretty much doing the same thing. Need to dig into what the difference is.

Reviewed By: arthaud

Differential Revision: D33554939

fbshipit-source-id: c169b7d12c61e7fff5c8a76ad645243b37913239