Remove sources when popping from a dictionary typed variable
Summary:
When popping a key from a dictionary typed variable and the popped value flows from sources, we should remove the source associated with the key, during the forward analysis.

This diff only handle the simplest case. That is, `d.pop(s)` when `d` is a variable and `s` is a string literal.

Reviewed By: arthaud

Differential Revision: D33540428

fbshipit-source-id: 696e3e08f6fdf5844e0fc76601c96a95c63cec53