Move handling of coverage requests to PyreQueryHandler
Summary:
# Context

See diff D33444367 (https://github.com/facebook/pyre-check/commit/a1882b7bdcf8389da12a2a1507473913cd22207e).

# This diff

In prep of querying the binary for if a module if typechecked at all or not; split the handling of `textDocument/typeCoverage` request between PyreServer and PyreQueryHandler. Similar to `types` query, we save it in the queue for PyreQueryHandler to respond.

For ease of review, this diff doesn't query the binary, just refactoring. That will follow separately. Also, gating of hover is moved from starting the query handler itself to responding to types queries.

Reviewed By: grievejia

Differential Revision: D33518093

fbshipit-source-id: 492efd2920e47f0b3e14c6432fe8fb6e93271a94