Add sinks when popping from a dictionary typed variable
Summary:
When popping a key from a dictionary typed variable and the popped value flows into sinks, we should add the sink associated with the key, during backward analysis.

This diff only handle the simplest case. That is, `d.pop(s)` when `d` is a variable and `s` is a string literal.

Reviewed By: arthaud

Differential Revision: D33566711

fbshipit-source-id: 6f2b632deda5d795805d902d173abc415a5f3632