Separate cache invalidation from the parsing logic
Summary:
As a follow-up of D33557517, we want to move away from using the hash from the `Metadata.t` of the parser.
There is no reason to have the cache invalidation logic tied to the parser logic.
It actually creates discrepancies when a parse error occurs, hence the hash is not stored even though the file exist.
Using a separate shared memory table to store the hash solves that problem.

In a follow-up, we will completely remove the hash from `Metadata.t` since pyre doesn't use that anymore.

Reviewed By: grievejia

Differential Revision: D33698219

fbshipit-source-id: e1bf2fd319c6da3f599a1258795ab0ce50761bb1