Clean-up the caching API
Summary:
This diff cleans up the cache implementation for pysa.

The previous implementation was using an imperative style which is hurting readability. For instance, we would have calls to `load` and `store` in the middle of the call graph building logic. This diff fixes it by implementing a cache API that takes a function that is called if the cache is absent.

The previous implementation would use exceptions and propagate them, making it hard to follow. This diff translates exceptions into `Result.t` which makes it easier to handle and prevents missing try/catch blocks.

Another benefit of this new implementation is that we do not need to remove any files when the cache is invalidated.

Reviewed By: dkgi

Differential Revision: D33723008

fbshipit-source-id: b0062d1e13dd80a3b639a0e3672b4a64777c9a1d