Update compare models script
Summary:
It seems like both the `"leaves"` and the `"kinds"` keys appear in the model output JSON (from D32146642 (https://github.com/facebook/pyre-check/commit/6c3279e6de6eff97ec69cf803bced106f6298da5)), so let's look at both of these keys when comparing models.

This came up when working on D33816152. For example, in P476718351 the `"leaves"` key is still in the JSON, so we should probably account for it.

Reviewed By: grievejia

Differential Revision: D33816159

fbshipit-source-id: 02278b557ceff9dbc37b7a406c9961b2edbc2d06