Remove callable wrapper for initialized models.
Summary:
Problem: For the earlier approach to decorator inlining, we needed to fetch part of the initial models before inlining and part after inlining. So, we needed a complicate `InitializedModels.t` data type representing a callable for the second computation.

We no longer need that with the new approach where we inline as part of type checking.

This diff removes the extra call to `get_models_including_generated_models`, which was a no-op. Also made `FixpointAnalysis.initialize_models` just return the models instead of returning a callable.

I've kept the `InitializedModels.t` module as such right now to keep this diff simple. Will remove the module in the next diff.

Reviewed By: arthaud

Differential Revision: D33856787

fbshipit-source-id: b7952e535b282849773721c838d67a5652b2eb47