Perform via-value and via-type expansion during trace expansion
Summary:
`ViaValueOf` and `ViaTypeOf` features are currently expanded (i.e replaced with the actual value or type) before trace expansion, right after reading the callee model from shared memory. This sometimes leads to a performance bottleneck because on big models we basically end up copying the entire taint tree.

One solution is to move via-value and via-type expansion during trace expansion, i.e `Taint.apply_call`. This function transforms the call information from the caller to the callee, which means taint from different callees are merged. By applying via-value and via-type expansion at the same time, we avoid the bottleneck and benefit from merging.

This also has other advantages. For instance, we only perform trace expansion for sinks that actually match a parameter. For instance, you could have a sink on an optional parameter. If the parameter is not specified at the call site, we now won't spend time expanding ViaValueOf and ViaTypeOf for it.

Finally, this means `via-type-of` and `via-value-of` will now be shown as a local breadcrumbs on the relevant frame, which helps triaging issues. Before, it was directly added to the propagated features hence it would only show up in the final set of breadcrumbs.

Reviewed By: dkgi

Differential Revision: D33832235

fbshipit-source-id: f3cdf5c51d69d949860b5704c69c5633fd47c940