False negative when taint flows into constructor of DataClass with tainted attribute
Summary:
The constructor of any dataclass assigns the actual arguments to the attributes, which is however invisible to pysa (due to being auto-generated).
Hence, when a dataclass has an attribute marked as a sink, there exists a flow from the actual arguments to the attributes, which is currently not reported by the analysis.

This diff fixes the above false negative.

Reviewed By: arthaud

Differential Revision: D33750088

fbshipit-source-id: b4582c6d8c2124a2953380fe2196e41bd8c34f27