Expose a potential bug in call graph construction by marking sinks on attributes of named tuples
Summary:
This diff was originally intended to address a false negative when taint flows into constructor of NamedTuple with tainted attribute. See below for details. However, the above false negative does not actually exist, because the preprocessing step automatically generates assignments from parameters of `__init__` methods to attributes, unlike dataclass.

The constructor of any named tuple assigns the actual arguments to the attributes, which is however invisible to pysa (due to being auto-generated).
Hence, when a named tuple has an attribute marked as a sink, there exists a flow from the actual arguments to the attributes, which is currently not reported by the analysis.
This diff fixes the above false negative. This issue is similar for a dataclass. See T106922147.

But why we still need this diff? A new test (that is relevant with named tuples) shows that there is a bug that may be caused by call graph construction, which is exposed by this diff.

Reviewed By: arthaud

Differential Revision: D33822823

fbshipit-source-id: 49a3f487216a907d11daa9750ae03bffa9eef7f2