Pass taint configuration errors from the binary to the client as json (2/3)
Summary:
This diff cleans up the error handling logic in pysa.
We used to raise exceptions to report errors in the taint configuration files. This diff converts errors into json that is read in the client and printed in the excepted format.
This diff also adds different binary error code for taint configuration errors and model verification errors. We also use a different client error code now, which will allow users of pyre to differentiate between type/configuration/model errors.

Reviewed By: dkgi

Differential Revision: D33797879

fbshipit-source-id: b2e4e59e9fbe27d225706111514f2c39085cf0e7