Catch and warn on failure to parse line break code block in pyre-upgrade
Summary:
Ran into crashing error suppression jobs while running the pyre infer codemods.

Turns out we aren't always passing a parseable block of code into the logic that handles adding error suppressions to line-broken code.

First let's add handling to catch the libcst exception and warn explicitly that the issue had to do with line breaks, and skip the file rather than crashing the entire upgrade job.

I'll follow up with a fix to check for open parens on a trailing line in a line-broken block of code, which can extend the block and is something we aren't currently handling.

Reviewed By: dkgi

Differential Revision: D33901118

fbshipit-source-id: d17b269df07f5083743a01c312e3cc0599f663b7