Avoid consolidating nested configurations that are properly ignored by parent
Summary:
Our autofix job tried to remove an intentionally nested local configuration that was properly ignored by the parent config: https://www.internalfb.com/diff/D33908608?transaction_fbid=1308941986246454

The autofix job is triggered when there is a user error (ie., pyre cannot run / cannot produce type errors, but it's not an infra problem like a timeout) and attempts to delete buck targets that are broken or have been removed, and also tries to delete nested configs.

It doesn't auto-land any of this, just puts up the diff and attaches it to the task to make it easier for whichever oncall is handling it.

That said, we should not even attempt the nested config step if we see that the parent local config is ignoring the subdirectory, which is a healthy setup.

Reviewed By: dkgi

Differential Revision: D33935240

fbshipit-source-id: 99e5906d1a778cc43c1775f20cd81d3e0da1cdd4