Rename the `Flow` module into `Issue`
Summary:
This might be a personal taste, but I feel like `flow` is too generic as a name.
I feel like `Issue` is a much better module name because, well, this module defines the final issue type that we produce.
In a follow-up, we will add submodules for all the types defines in this module, i.e `flow`, `candidate`, `issue`, to avoid name clashes.

Reviewed By: dkgi

Differential Revision: D33961810

fbshipit-source-id: c209556f1484a49a5cfd157f0ef55c0d3e7323ab