Compute breadcrumbs of issues when exporting to json
Summary:
There is no real need to store the set of breadcrumbs of an issue in a `Flow.issue` since it can easily be computed from the `flow`. Let's save some space and time by computing it when we dump the issue to json.
This will also make my work easier.

Reviewed By: dkgi

Differential Revision: D33961050

fbshipit-source-id: 50ec45e0ad889acd340d180f53f9e2bbf43697e4