TestHelper: specify tito_parameters in full
Summary:
# This diff

Noticed that in the tests for matching tito_parameters, tito_parameters outcome is specified in somewhat of a hacky way, and relying on string matching. Instead of continuing/extend this string-based matching as we will be adding support for transforms in tito, clean this up so we match it structurally instead.

For instance, for model `arg: TaintInTaintOut[Updates[self]]` the outcome was being specified as a *string* with contents `arg updates parameter 0`. Refactor to specify it as `{name: "arg"; sinks: [Sinks.ParameterUpdate 0]}` instead.

In addition to be cleaner, we are also testing the expected behavior more precisely in many pre-existing tests (e.g. for Attach we weren't actually checking that it is an Attach).

Note this diff doesn't introduce any new tests / any support for transform; just a refactoring diff for pre-existing tests.

# Context

Not needed for this diff, but broader context on taint transforms linked in D33831318 (https://github.com/facebook/pyre-check/commit/3f71b3c25bb63bd09fd71455dedc67d13405ef5a).

Reviewed By: grievejia

Differential Revision: D34013128

fbshipit-source-id: 4da1feac797a1354604a862fa676f7bc356a5e4e