Add a dummy call index in the call graph [issue dedup 1/n]
Summary:
# Problem
To identify unique issues across runs, sapp uses a issue handle composed of the root callable, the rule code, the relative line number and column number where sources meets sinks (more precisely, the position of the sink). The problem is that a small line change between two runs will triggers different issue handles, and hence different issues, leading to a lot of duplicates.

# Solution
Instead of using the line number and column number, let's use the name of the callee leading to the sink, the argument name, as well as the textual order index of the callee in the call graph.
This is robust to line changes. The downside is that there is a small risk of ignoring new issues when adding a call to a function before an existing call to the same function, as this will mess up indices.

# Implementation
In this diff, we are introducing a dummy *textual order index* stored in the call graph. We set it to zero for now.

Reviewed By: dkgi

Differential Revision: D33992272

fbshipit-source-id: 2c7a8f63526d1463592e0ce176ff6740aba800e1