Introduce a handle that uniquely represents a sink [issue dedup 3/n]
Summary:
See previous diffs for context about the issue deduplication problem.

This diff introduces `SinkHandle.t`, which uniquely represents the location of a sink in a callable, and is robust to line changes.

For sinks on calls, this is composed of the fully qualified name of the callee, the textual order index of the call, and the root of the callee port.
For instance:
```
def sink_with_hop(x=None, y=None, z=None): sink(y)
def f():
  a = source()
  sink_with_hop()
  sink_with_hop(z=0, x=1, y=a) # handle is `sink_with_hop|1|y`
```
For sinks on attributes, this is the fully qualified name of the attribute and the textual order index of the attribute (i.e, N if this is the N-th access to this attribute within the callable).
For return sinks, this is the placeholder `return`. In the future, we could potentially add an index as well (i.e N-th return statement in the function).
For string literal sinks and conditional test sinks, this is simply the sink kind.

Note that this is still unused. See the follow-up.

Reviewed By: dkgi

Differential Revision: D34022798

fbshipit-source-id: c1f6931372f65eb92481dced52b6817c5ec9534c