Move the issue grouping logic in the issue module
Summary:
Right now, we have some of the issue grouping logic present in the forward analysis.
In the future, we will want to change that logic to group issues with the new proposed handle.
To make that easier, let's move the grouping logic from the forward analysis to the issue module, and make a clean API for it, named `Candidates`.

Reviewed By: dkgi

Differential Revision: D34098399

fbshipit-source-id: aa026703f0ec157ff69b2fb30ab918f98e5bd306