Migrate fixme threshold commands to common API
Summary:
This diff migrates several commands that deal with gathering and suppressing errors within a specified fixme threshold to use the common API rather than re-rolling the logic each time.

It moves some of the relevant tests out of the specific command and into the filesystem test file where the functionality actually lives.

This fixes a bug where the error source would not be respected (errors will always be regenerated) if a fixme threshold is passed into fixme-single.

Reviewed By: pradeep90

Differential Revision: D34082263

fbshipit-source-id: 55e9184466156f3ec0736ac95b60a5055a2cf2b9