Refactor support for fixme threshold into general error suppression logic
Summary:
Found some bugs in our error suppression handling on the sandcastle side that duplicates the work generating errors.

This is a nfc lead-up to the fixes, where we make the error suppression common method support a fixme threshold that is used in many upgrade commands. The logic is currently duplicated (and buggy) across a lot of these. Also add some testing.

Reviewed By: pradeep90

Differential Revision: D34063010

fbshipit-source-id: 15970f3e9ae5a76c4ba7273bf576d721117643ce