Handle suppressions on multi-line string close
Summary:
Ran into issues during the inference codemod across fbcode where we'd break the AST when suppressing errors on the same line as a multi-line string close.

For example, `schedulers/chronos` project:

```
    def testChronosRunningOnHostJobName(self) -> None:
        sys.stdout = StringIO()
        self.assertEqual(
            0,
            main([
                "running", '-c', 'test',
                '--fields', 'jobInstanceId,job.name', 'chronos041.prn2']))
        self.assertEqual("""\
1\ta
2\tb
3\tc
# pyre-fixme[16]: `TextIO` has no attribute `getvalue`.
""", sys.stdout.getvalue())
```

Let's keep this simple by sticking the fixme on the end of the line if the line closes a multi-line string rather than try to deal with re-formatting this situation.

Reviewed By: pradeep90

Differential Revision: D34012033

fbshipit-source-id: 7f591b06579fa9aed5750acd2709aa48c978331e