Quote type annotations that are not defined at runtime in pyre-infer
Summary:
By far the biggest reason many inference codemod diffs could not land was the introduction of type annotations that weren't defined at runtime because they referenced method or attribute's parent class.

Since "quote_annotations" is a command-wide option that quotes the entire qualified annotation, let's add a different flag denoting the runtime evaluability of a specific annotation, and quote the sanitized version in `to_stub` if it is not defined at runtime due to referring to a parent.

Reviewed By: dkgi

Differential Revision: D34116052

fbshipit-source-id: 8c5c4fc4594741cc6f52e34055f4c1e48db64bc6