Avoid adding annotation declarations for unnamed variables
Summary:
This was another issue with pyre infer that resulted in some abandoned diffs in the first codemod pass through fbcode.

If there is a multi-line assignment, we need to pull type declarations up onto their own line. The existing codemod also adds annotations for unused `_` variables, ie:

```
_: str
_filename: str

_, _, _filename = __name__.rpartition(".")
```

example codemod diff: https://www.internalfb.com/diff/D33920169

We should avoid adding this noise - I think it's OK that the pyre infer backend doesn't check for variable name, and does return the annotation of a variable named `_` in the inference JSON results if it exists. However, when we actually go to apply type stubs to the code in place, if we see a multi-line string we shouldn't pull out and annotate `_` variables on their own line.

Note: if the original code says `_ = 1 + 1` then we /will/ still annotate it to be `_: int = 1 + 1`, which I think is good.

Reviewed By: dkgi

Differential Revision: D34123841

fbshipit-source-id: a483fde790d0134a248323cdca554b5f641f6e29