Remove extra logic to log and reraise exceptions
Summary:
I'm honestly not sure why this was introduced in the first place (probably for debugging?).
The model parser should not raise exceptions anymore, but just return `Result.t`.
It seems safe to remove this extra logic. In the worst case, exceptions should be logged anyway at the outermost layer of the binary.

Reviewed By: pradeep90

Differential Revision: D33798593

fbshipit-source-id: 201b657f9ea448631a2860579580a0dd7f792dce