Make `pyre coverage` take a `--path` option instead of variadic paths.
Summary:
Problem: I want to add a command for calculating expression-level type coverage. `pyre coverage expression-type` seems like a natural home. This will make it easier to run from a sandcastle job compared to a script that has to be deployed separately. We could make `pyre coverage type-annotation` do the current type-annotation calculation, with `pyre coverage` defaulting to `pyre coverage type-annotation`.

However, `pyre coverage` currently takes a variadic `roots` argument. This means it's ambiguous whether `pyre coverage expression-type` is calling a subcommand or passing a root. `click` treats it as a root.

RFC on the plan in this diff:
1. Add a repeated option `--root` so that we can do `pyre coverage --root root1 --root root2` instead of `pyre coverage root1 root2`. Release the client.
2. Make Sandcastle call `pyre coverage` using `--root`.
3. Remove the old roots argument and add subcommands `pyre coverage type-annotation` and `pyre coverage expression-type`.

The downside of `--root` is that it can make CLI usage more cumbersome. We can't do `pyre coverage a b c d`. We have to do `pyre coverage --root a --root b --root c --root d`. If that's fine, I think it's the best place for the expression-level coverage script.

Alternative designs:
+ Add a new command `pyre expression-coverage`. It adds yet another command to pyre, but I'm ok with it since it keeps things simple.
+ Add a flag `pyre coverage --expression-level ...` that computes expression-level coverage. This forces the two types of coverage to use the same CLI arguments.
+ Just add a script. This would be harder to run from Sandcastle since we have to set up an fbpkg for it. It would also be harder to discover.

Note that `pyre statistics` has a similar repeated argument, so we can't move expression-level coverage there.

Reviewed By: grievejia

Differential Revision: D34148528

fbshipit-source-id: 1d6dc9b9d279d40fe5d2fe49b9e38ae2ffca3115