Propagate taint transforms in backward analysis (sinks and tito)
Summary:
# This diff

Follow-up on D34022496 (https://github.com/facebook/pyre-check/commit/2c8cbdece76df12200c13970ddff680d94866a75), backward propagation of taint transforms is also relatively straightforward, except for the note below.

**On the order of transform in TITO**
So far since our model parser syntax didn't support parsing a sequence of TITO transforms in the same model, and we weren't propagating transforms in TITO models, the length of sequence of transform in a tito was at most one. With TITO propagation, we can end up with longer sequences, so we need to pick a convention, and account for that when applying. Since we store TITO as part of Sinks, it is natural to pick the same ordering as in Sinks. Thus, we need to the `reverse` when applying in sources. Those are precisely the changes introduced in `sources.ml` in this diff.

# Context

See pointers in D33831318 (https://github.com/facebook/pyre-check/commit/3f71b3c25bb63bd09fd71455dedc67d13405ef5a).

Reviewed By: arthaud

Differential Revision: D34136192

fbshipit-source-id: 2dc6669ab87132c3d57985d559e8a1110e29c52b