Unify querying in query handler (1/2)
Summary:
# Context

Follow-up from D33676393. Server connection logic in the query handler was getting duplicated.

# This diff

Deduplicate it by unifying the querying logic a bit. There is still some duplication in interpreting the response, which is deduplicated in the follow up diff.

The socket_path computation is fixed based on the server_start_options, so we don't need to keep passing server_start_options around, and simply upfront compute and pass the socket_path instead.

Reviewed By: grievejia

Differential Revision: D33745893

fbshipit-source-id: ae86c75aeeca2adb08c6a46d92857736f590cf1f