Make ModulesOfPath query to detect if a file is typechecked
Summary:
# Context

See diff D33444367 (https://github.com/facebook/pyre-check/commit/a1882b7bdcf8389da12a2a1507473913cd22207e).

# This diff

Query the binary for if a module if type-checked at all or not.

To-follow: there is some cleanup possible of "excludes" in persistent that we no longer need to track. Will do the cleanup in a separate diff, so as to keep this diff with more semantic changes smaller.

Reviewed By: grievejia

Differential Revision: D33676393

fbshipit-source-id: f6866a6d68e4f0d2c2dfd8428d53101743bd6353