Add pysa_version to .pyre_configuration (1/n)
Summary:
# Motivation

Currently the pysa client version is behind a sitevar `PYRE_PYSA_VERSION_HASH`. With it being a sitevar, upgrades to the version require the test results to be copied to the test plan D34045248 and make it hard to pin down a specific version when debugging.

# Implementation

To facilitate a smooth migration away from the sitevar, this diff will allow a `pysa_version` to be specified in the `.pyre_configuration` file, the next set of diffs will add a `pysa_version` to `.pyre_configuration` in fbcode and distillery, and then we can add logic to use that version over the sitevar

Reviewed By: arthaud

Differential Revision: D34145168

fbshipit-source-id: 938b84b2f46a53a550b9efcb4373c5f7c9d87cad