Model/annotation parsing: seperate out parse_tito
Summary:
# This diff

Even though we are using the same data-structure internally in our representation (Sinks.t), tito are a separate concept from sink -- separate these out for parsing purposes, which also allows for better error handling.

Some of the test changes are because we are catching errors earlier, with different messages.

# Context

Locally: separating out a refactoring-only diff in the process of working on better error handling with TITO transforms in models.

Reviewed By: arthaud

Differential Revision: D34153464

fbshipit-source-id: 86c49ca2c1fea12fb13cd7837f2f3eea5236680d