Fix OSS build: unittest.mock is more limited in < 3.8
Summary:
We recently added unit test code that reaches into a MagicMock's
`__aenter__` attribute, and treats a mock as always a valid
context manager.

This works in Python 3.8+, but not in 3.6 and 3.7 so our OSS build
broke. In this commit, I instead set the return value to a manually
written async context manager.

I verified locally that this fix works but I'll also export to github
and check that we get clean results before landing.

Reviewed By: kbansal

Differential Revision: D34275880

fbshipit-source-id: 59dd4583e52c70a533a14a4ada5486aa528ed2a5