Update playground to use incremental.
At some point in 2021, there was a problem where pyre
incremental did not work on Ubuntu so the server was updated to use
`pyre check`.

This is a major performance issue, and moreover at this point it
appears that pyre works just fine in an ubuntu server. So, this
commit updates the code to:
- initialize the pyre server for a given worker prior to configuring
  routes (I don't want the server to be up and not initialized - that
  won't give great behavior when we rotate hosts, given that each
  worker would have to be hit with a request).
- use pyre incremental in both initialization and the check request