Create callees in the call graph for implicit calls to str and repr
Summary:
# What does this diff do?

For any f string `f"prefix{e}suffix"`, this diff lets forward and backward analysis to analyze (implicit) calls to `str(e)` (see [this](https://docs.python.org/3/library/string.html#formatspec)), instead of `e`.
Furthermore, `str(e)` will be resolved as either `e.__str__` or `e.__repr__` depending on whether the value of `e` defines method `__str__` (see [this](https://docs.python.org/3/library/stdtypes.html#textseq)).

To realize the above logic, we add calls to `__str__` or `__repr__` for f strings in the call graph, such that the location of `e` is mapped to calls to either `__str__` or `__repr__` in the call graph, which may have models.

# Why need this diff?

After this diff, we should be able to detect flows that were missed due to not handling implicit (not explicit) calls to `str` (such as an f string `f"{x}"` that implicitly calls `str(x)`), thus fixing a class of false negatives.

Reviewed By: arthaud

Differential Revision: D34043995

fbshipit-source-id: f712910c7c0a18a8137c2a315385a99f0915d584