Avoid removing existing starred arg annotations in infer
Summary:
Noticed during inference codemods that pre-existing annotations on starred args were being deleted for no apparent reason.

It turns out I think there is a bug in the annotation application transform, where rather than taking the existing node and making changes to the pieces where we want to set annotations, we're taking the inferred annotations as the base object and returning that.

cc stroxler are we tracking delta of this module vs. the open sourced version of LibCST? What's the best way to keep changes here maintainable (I think I've made one or two over the course of bug-correcting during the codemods)?

Reviewed By: stroxler

Differential Revision: D34277896

fbshipit-source-id: 3d7719711bc0edfd68bb0e9e98f5094f062bdb7a