Preserve existing type information in local inference
Summary:
At the risk of opening a bigger can of worms, my investigation into why we inference codemod diffs needed to be abandond bc weren't accounting for the types of parameter default values (see context: D34161066 (https://github.com/facebook/pyre-check/commit/787e576f124f2992e0e848b0a0f167e05338aed0)) has uncovered some broader questions about how we structure the inference fixpoint.

General structure of pyre inference type propagation, for context:
1. We run TypeCheck's initialize state to get a starting set of local annotations from the define signature
2. We run Infer's `initial_forward` modifier to reset some `Any` annotations back to `Bottom` so we don't lose type information we gather in the subsequent pass
3. We run Infer's `forward`, which special cases some unassigned call expressions, etc., but primarily just shells out to Typecheck's `forward` to propagate type information.
4. We pass the resulting state into Infer's `initial_backward` to add type information for a dummy `$return$` local variable from the return annotation if it exists, something normal type check can ignore
5. We run Infer's `backward` which special cases some assign statements and return statements to propagate type info
6. We continue the forward/backward cycle until a fixpoint is reached.

The issue at hand here is that we aren't actually gathering all possible types that a local variable might have over the course of the function to suggest an annotation.
There are some forward-specific and backward-specific assumptions we're making that may make sense in a forward-only or backward-only world where we want correctness at each point of the control flow, but don't really make sense in the combined inference setting here where we recommend one annotation for the local variable overall.

**Forward loses info (not addressed in this diff):**

```
def foo(x = 1) -> None:
   x = "string"
```

Will annotate the parameter `x` as having type `str`.

This is because for the purposes of forward type propagation, if there isn't an immutable type given to a value, all we care about is the most recent mutable value assigned to a local. It makes perfect sense to throw away the previous type information as we learn new type information.

However, for type inference purposes, as long as we are allowing local variables to change their type over the course of a function scope, shouldn't we want to suggest the type annotation that can accommodate all of the types that local takes on at any point in the function? ie., stop throwing away the previous annotation info and join it in instead?

**Backward loses info (addressed in this diff):**

```
def foo(y = 1) -> str:
   y = unknown()
   x = y
   return x
```

Will annotate the parameter `y` as having type `str` despite it clearly possibly sometimes being an `int`.

Side note: If we really wanted to be precise about propagating types backward, seeing an assignment like `x = y` should actually completely wipe everything we know about `x` since we are going bottom-up. `x` could be literally anything before such an assignment, and our post-assignment state would not be affected.

However, we don't wipe this info, precisely because we are OK making some theoretical correctness tradeoffs for retaining information that will be relevant to a suggested annotation the vast majority of the time in the wild.

This doesn't need to be extrapolated much further to say that we probably should be OK with collecting type information across the local scope of the variable to make our type inference suggestion the most useful and landable, rather than wiping/dropping information that is specific to forward-only or backward-only per-statement correctness.

 ---

Thoughts/concerns welcome - so far this stack doesn't touch forward logic yet, so it won't fix the original common issue during the codemod pass where `None` is often used as a default param value and we don't make the suggested annotation optional:

```
def foo(x: int = None) -> None:
   ...
   x = 1
   ...
```
I might not prioritize fixing all of this logic right away, but since this can of worms was opened figured this is a high level choice we should iron out.

Reviewed By: dkgi

Differential Revision: D34161077

fbshipit-source-id: 7ad64c103f5fd2cf55620262e535ce118d609279