Use signature selection logic to match args and params in infer
Summary:
yikes_hard

It turns out we had some hand-rolled logic in the inference backend that did an extremely naive fold over arguments and parameters to propagate types.

That meant we were getting things wrong in the most basic case where args aren't being matched exactly in order:

```
      def foo(a: int, b: str) -> None:
          pass

      def bar(a, b) -> None:
        foo(b=b, a=a)
```

I caught this when skimming a pyre infer codemod diff, and hope that I haven't missed too many instances where we land a bad type like this.

This diff re-uses signature selection logic to match args and parameters properly, and uses that mapping to propagate types when the argument is a simple name.

Reviewed By: stroxler

Differential Revision: D34275474

fbshipit-source-id: 95f14507c5ce16b4079ca769778db54f4a62d1fb