Wipe backwards resolution on assigns but save annotation snapshot
Summary:
This diff extends the `snapshot_resolution` introduced in D34292909 to support joining in default parameter types after inference is over, to also support using `meet` in a world where we care about inferring the overall type of a variable over the course of a local scope.

We need `meet` in the backwards pass in order to infer the broadest type that still works for all usages of the variable in a function. For example:

```
def foo(x):
   takes_int(x)
   takes_float(x)
```

Here, `x` must be the type of `meet int float = int`. If we did not do a meet of all the usages, we might infer that `x` is a float, which is not true and throws errors on the first call to `takes_int`.

However, `meet` starts to fall apart when we run into assignments:

```
def foo(x):
   x = 1
   takes_int(x)
   x = 1.0
   takes_float(x)
```

Here, the annotation we infer for `x` actually needs to be `float`. In fact, inferring `int` would create a type error because it invalidates the assignment happening on the third line.

Same idea but even more egregious:

```
def foo(x):
   x = 1
   takes_int(x)
   x = "string"
   takes_str(x)
```

Inference logic would go through the forward pass and join 1 and "string" to infer `int | str` by the exit state, then take this and go backwards and meet this union type with `str` on the first call it sees, resolving to `str`, then meet this with `int` on the second call it sees, resulting in `Type.Bottom` and no inference.

It's worth noting here that backwards inference is correct that at the line `takes_str`, `x` is not in fact `int | str`. It definitively is `str` only. When we see an assignment to `x`, correct backwards logic should in fact wipe everything we know about `x` above that assignment, as `x` could have been anything above the assign and still be consistent with our pre-existing knowledge below it.

Historically we declined to wipe our inference on encountering an assign in the backwards pass (and also didn't do meets properly) because we didn't want to lose all knowledge that `x` was a `str` at some point in this function.

However, if we follow the thought experiment and say that we are implementing backwards inference correctly, we should think that on line 4, `x` is a str; on line 3, we wipe everything we know about `x`, and on line 2, we think that `x` is an int, and on line 1, we wipe everything we know about `x`.

This paints a completely accurate picture of the inferred type of `x` at each point of the logic in the function body. However, what we want at the end of the day is a single annotation we can attach to `x` that accounts for all of the types that it might be at some point in the function body.

----

This diff implements backwards inference "correctly" for the first time - using `meets` whenever we infer the type of a usage, and wiping everything we know from the resolution when we see an assign.

The trick is that when we wipe everything we know, we'll save a snapshot of the current inferred type in `snapshot_resolution`, which (along with the default parameter value type, if it exists) will be joined together at the end to come up with the inferred explicit annotation we want to give to the variable.

This means we can use `meet` to get an actually accurate type at each point in the function, but can also provide an explicit annotation that encompasses types getting broader or narrower through explicit assigns - aka. we can get the right annotation on all examples above.

Reviewed By: dkgi

Differential Revision: D34305388

fbshipit-source-id: 2fe0d0dd23f66632885d744f053c6c051032619f